+++ src/nsidc/metgen/constants.py
@@ -35,6 +35,7 @@ SETTINGS_SECTION_NAME = "Settings"
 
 # File name defaults
 PREMET_SUFFIX = ".premet"
+SPATIAL_SUFFIX = ".spatial"
 CSV_SUFFIX = ".csv"
 NETCDF_SUFFIX = ".nc"