+++ src/nsidc/metgen/metgen.py
@@ -24,7 +24,16 @@ from typing import Optional
 import earthaccess
 import jsonschema
 from earthaccess.exceptions import LoginAttemptFailure, LoginStrategyUnavailable
-from funcy import all, concat, filter, first, notnone, partial, rcompose, take
+from funcy import (
+    all,
+    concat,
+    filter,
+    first,
+    notnone,
+    partial,
+    rcompose,
+    take,
+)
 from jsonschema.exceptions import ValidationError
 from pyfiglet import Figlet
 from returns.maybe import Maybe
@@ -234,11 +243,14 @@ class Granule:
     data_filenames: set[str] = dataclasses.field(default_factory=set)
     browse_filenames: set[str] = dataclasses.field(default_factory=set)
     premet_filename: Maybe[str] = Maybe.empty
+    spatial_filename: Maybe[str] = Maybe.empty
     ummg_filename: Maybe[str] = Maybe.empty
     submission_time: Maybe[str] = Maybe.empty
     uuid: Maybe[str] = Maybe.empty
     cnm_message: Maybe[str] = Maybe.empty
-    data_reader: Callable[[str, config.Config], dict] = lambda auth_id, cfg: dict()
+    data_reader: Callable[[str, str, str, config.Config], dict] = (
+        lambda auth_id, cfg: dict()
+    )
 
 
 @dataclasses.dataclass
@@ -316,7 +328,7 @@ def process(configuration: config.Config) -> None:
 
 def data_reader(
     auth_id: str, data_files: set[str]
-) -> Callable[[str, config.Config], dict]:
+) -> Callable[[str, str, str, config.Config], dict]:
     """
     Determine which file reader to use for the given data files. This currently
     is limited to handling one data file type (and one reader) per collection.
@@ -444,19 +456,19 @@ def validate_cmr_response(umm: list):
     return umm[0]["umm"]
 
 
-def ummc_content(umm: dict, key: str):
+def ummc_content(ummc: dict, key: str) -> str:
     """
     Look for a key in a UMM-C record (or a piece of the record), and log the status.
     """
     val = None
     logger = logging.getLogger(constants.ROOT_LOGGER)
 
-    if umm is None:
+    if ummc is None:
         return val
 
     try:
-        val = umm[key]
-        logger.info(f"{key} information in umm-c response from CMR: {val}")
+        val = ummc[key]
+        logger.debug(f"{key} information in umm-c response from CMR: {val}")
     except KeyError:
         logger.info(f"No {key} information in umm-c response from CMR.")
 
@@ -524,7 +536,12 @@ def grouped_granule_files(configuration: config.Config) -> list[tuple]:
     Identify file(s) related to each granule.
     """
     file_list = [p for p in Path(configuration.data_dir).glob("*")]
-    premet_file_list = premet_files(configuration)
+    premet_file_list = ancillary_files(
+        configuration.premet_dir, constants.PREMET_SUFFIX
+    )
+    spatial_file_list = ancillary_files(
+        configuration.spatial_dir, constants.SPATIAL_SUFFIX
+    )
 
     return [
         granule_tuple(
@@ -533,23 +550,19 @@ def grouped_granule_files(configuration: config.Config) -> list[tuple]:
             configuration.browse_regex,
             file_list,
             premet_file_list,
+            spatial_file_list,
         )
         for granule_key in granule_keys(configuration, file_list)
     ]
 
 
-def premet_files(configuration: config.Config) -> list[Path]:
-    if configuration.premet_dir:
-        premets = [
-            p
-            for p in Path(configuration.premet_dir).glob(f"*{constants.PREMET_SUFFIX}")
-        ]
-        if not premets:
-            raise Exception(
-                f"Premet directory {configuration.premet_dir} is empty or unreadable."
-            )
+def ancillary_files(dir, suffix):
+    if dir:
+        files = [p for p in Path(dir).glob(f"*{suffix}")]
+        if not files:
+            raise Exception(f"Directory {dir} is empty or unreadable.")
 
-        return premets
+        return files
 
     return None
 
@@ -591,9 +604,10 @@ def granule_tuple(
     browse_regex: str,
     file_list: list,
     premet_list: list,
+    spatial_list: list,
 ) -> tuple:
     """
-    Important! granule_regex argument must include a captured match group
+    Important! granule_regex argument must include a captured match group.
 
     Return a tuple representing a granule:
         - A string used as the "identifier" (in UMMG output) and "name" (in CNM output).
@@ -613,23 +627,28 @@ def granule_tuple(
         str(file) for file in file_list if re.search(granule_key, file.name)
     } - browse_file_paths
 
-    if premet_list is None:
-        premet_file = None
-    else:
-        premet_matches = [
-            str(file) for file in premet_list if re.search(granule_key, file.name)
-        ]
-        if not premet_matches:
-            premet_file = ""
-        else:
-            premet_file = first(premet_matches)
-
     return (
         derived_granule_name(granule_regex, data_file_paths),
         data_file_paths,
         browse_file_paths,
-        premet_file,
+        matched_ancillary_file(granule_key, premet_list),
+        matched_ancillary_file(granule_key, spatial_list),
     )
+
+
+def matched_ancillary_file(granule_key, file_list):
+    if file_list is None:
+        file_match = None
+    else:
+        file_matches = [
+            str(file) for file in file_list if re.search(granule_key, file.name)
+        ]
+        if not file_matches:
+            file_match = ""
+        else:
+            file_match = first(file_matches)
+
+    return file_match
 
 
 def derived_granule_name(granule_regex: str, data_file_paths: set) -> str:
@@ -708,7 +727,7 @@ def create_ummg(configuration: config.Config, granule: Granule) -> Granule:
     metadata_details = {}
     for data_file in granule.data_filenames:
         metadata_details[data_file] = granule.data_reader(
-            data_file, granule.premet_filename, configuration
+            data_file, granule.premet_filename, granule.spatial_filename, configuration
         )
 
     # Collapse information about (possibly) multiple files into a granule summary.