+++ CHANGELOG.md
@@ -3,9 +3,12 @@
 * Add support for reading IRWIS2 CSV files (Issue-154)
 * Store `GranuleSpatialRepresentation` value from CMR collection metadata in the
   `Collection` object.
-* Add option to read temporal extents from premet files. (Issue-160)
+* Read temporal extents from `premet` files if they exist. (Issue-160)
+* Read spatial information from `spatial` files if they exist. (Issue-161)
+* Note successful collection metadata harvest details in the log, not to `stdout`.
+* Move `netcdf_reader` to `readers` module.
 
-## v1.4.0rc7
+## v1.4.0
 
 * Add a CSV reader to process [SNOWEX](https://nsidc.org/data/snex23_ssa/versions/1) granules. (Issue-108)
 * Handle a bug in `xarray` when processing NSIDC-0630 v2 granules. (Issue-152)